List all docker containers when cleaning up

+++ docker/recipes/deploy.rb
@@ -21,13 +21,13 @@ node[:deploy].each do |application, deploy|
   bash "docker-cleanup" do
     user "root"
     code <<-EOH
-      if docker ps | grep #{deploy[:application]}; 
+      if docker ps -a | grep #{deploy[:application]};
       then
         docker stop #{deploy[:application]}
         sleep 3
         docker rm #{deploy[:application]}
       fi
-      if docker images | grep #{deploy[:application]}; 
+      if docker images | grep #{deploy[:application]};
       then
         docker rmi #{deploy[:application]}
       fi
